# docs/swagger.yaml
basePath: /
definitions:
  models.Ingredient:
    description: Modelo para gerenciamento de ingredientes.
    properties:
      id:
        description: ID é o identificador único do ingrediente.
        type: integer
      name:
        description: Name é o nome do ingrediente.
        example: Farinha de trigo.
        type: string
    type: object
...
host: localhost:3000
info:
  contact:
    email: sslindner@furb.br
    name: Sofia Sousa Lindner
  description: Crie suas próprias receitas e faça um incrível livro de receitas!
  license:
    name: MIT
    url: https://github.com/Marlon-Sbardelatti/go-rest-api/blob/master/LICENSE
  title: Cookbook API
  version: "1.0"
paths:
  /ingredient:
    get:
      description: Buscar todos os ingredientes cadastrados
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Ingredient'
            type: array
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Buscar todos os ingredientes
      tags:
      - ingredient
    post:
      consumes:
      - application/json
      description: Criar novo ingrediente
      parameters:
      - description: Novo ingrediente
        in: body
        name: ingredient
        required: true
        schema:
          $ref: '#/definitions/models.Ingredient'
      produces:
      - text/plain
      responses:
        "201":
          description: Ingredient created!
          schema:
            type: string
        "400":
          description: Invalid JSON
      security:
      - Token: []
      summary: Criar novo ingrediente
      tags:
      - ingredient
  /ingredient/{id}:
    delete:
      description: Deletar ingrediente pelo ID
      parameters:
      - description: ID do ingrediente
        in: path
        name: id
        required: true
        type: integer
      produces:
      - text/plain
      responses:
        "200":
          description: Ingredient deleted!
          schema:
            type: string
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      security:
      - Token: []
      summary: Deletar ingrediente
      tags:
      - ingredient
    get:
      description: Buscar ingrediente pelo ID
      parameters:
      - description: ID do ingrediente
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Ingredient'
            type: array
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Buscar ingrediente pelo ID
      tags:
      - ingredient
    put:
      consumes:
      - application/json
      description: Atualizar ingrediente pelo ID
      parameters:
      - description: ID do ingrediente
        in: path
        name: id
        required: true
        type: integer
      - description: Ingrediente atualizado
        in: body
        name: ingredient
        required: true
        schema:
          $ref: '#/definitions/models.Ingredient'
      produces:
      - text/plain
      responses:
        "200":
          description: Ingredient updated!
          schema:
            type: string
        "400":
          description: Invalid JSON
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      security:
      - Token: []
      summary: Atualizar ingrediente
      tags:
      - ingredient
  /ingredient/name/{name}:
    get:
      description: Buscar ingrediente pelo nome sem case sensitive e convertendo '-'
        para espaços
      parameters:
      - description: Nome do ingrediente
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Ingredient'
            type: array
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Buscar ingrediente pelo nome
      tags:
      - ingredient
...
securityDefinitions:
  Token:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"